Force realization of non-trivial RHS in parallel multiplication.
This ensures that the various coeff() calls inside the multiply methods
do not trigger reevaluation of arbitrarily complicated Eigen
expressions. Rather, they are evaluated once on input and then freely
accessed. The exception is for any expressions that support data() -
i.e., existing VectorXd objects or map/block views - which suggests that
the result is already in memory and can be accessed without extra work.