Require wrappers to implement workspaces for intermediate objects.
Specifically, these are intermediate objects that depend on the RHS in the
wrapper class's multiply or adjoint_multiply methods. We don't want to have to
re-allocate memory for these intermediates upon every call to multiply, so the
idea is to allocate it once in the workspace at the start of the calling
function and then re-use that memory across multiply calls. This means that the
wrapper contract is extended to define an (adjoint_)workspace method and the
associated classes; a bit tedious for third parties, but whatever.

The workspace object is also passed to each (adjoint_)multiply method.
This allows us to keep the wrappers themselves const, as the workspace
is mutated separately. This is nicer to ensure that we retain
thread-safety across the various methods after construction.

A consequence of this change is that the parallel sparse matrix also
needs to update its methods to generate a workspace and accept it in its
multiply methods. Note that the workspace here is a no-op placeholder.

To make life easier for third parties, we provide the WrappedWorkspace
class. This allows wrappers to define their own workspace classes that
contain the workspace for the wrapped matrix, without needing to worry
about the exact type of the wrapped matrix.