Switch to subpar::parallelize_simple() for parallelization.
We wrap this in an irlba::parallelize() function, which provides a central
(documented) location to control the parallelization scheme. It also provides
an easy mechanism to switch out with a custom macro function, which we can't
use directly with a lambda because of funky substitution issues.

This requires some more care and testing for the EigenThreadScope class because
there are multiple ways in which OpenMP could be enabled or disabled, e.g.,
via subpar's default or by irlba's custom overrides.