Use the JacobiSVD algorithm everywhere, for consistency.
This should be fast enough for small matrices. It should also be fine
for thin/fat matrices where it might be used as a fallback by compute();
its quadratic runtime just depends on the smaller dimension anwyay.

Also increase the rigor of the basic tests by turning down the
convergence tolerance so that IRLBA is forced to agree with SVD.
This allows us to check all singular triplets at a low error.