Overhauled matrix interfaces to support run-time polymorphism. (#19)
We introduce Matrix and Workspace interfaces that can be inherited from. Each
Matrix subclass produces instances of various Workspace subclasses, the latter
of which handles the actual multiplication/realization with non-const
intermediates. Eigen matrix instances are now wrapped in a SimpleMatrix to
avoid complicating the rest of the library with compile-time switches.

This change allows us to switch between the input matrix classes at runtime, to
avoid realizing a new instance of the compute() function for each input class.
Hopefully this should reduce the compilation time and binary size.  We still
provide a template option for users to specify the concrete type in compute()
at compile-time, and all subclasses are marked as final to assist
devirtualization. In fact, if the exact type is known at compile-time, a
non-virtual function will be called to perform the multiplication/realization.

The compute() functions now only accept a Matrix-compliant object. Eigen matrices
need to be supplied to compute_simple() instead. We also rename the PCA function
to pca(), which also accepts Eigen matrices.