Choose better integer types and protect against integer overflow. (#20)
This mostly affects compute() and ParallelSparseMatrix, where some of the
arithmetic and implicit casts for allocations were a bit too relaxed. We add a
dependency on the sanisizer library to ensure that these calculations are ok.

We improve the worksharing across threads in ParallelSparseMatrix, and we
refactor the primary starts/ends into a single vector to reduce memory usage.
This involves some renaming of the getters for the per-thread data structures.

Finally, we slapped const on everything to indicate immutability.