Renamed arguments and outputs of pca() for to be more PCA-relevant.
So that the users get the PC scores and the rotation matrix. We also compute
the variance explained by each PC, instead of just having the singular values.

Refactored the internals of pca() to a single call to compute(). This involves
some fixes to CenteredMatrix and ScaledMatrix to work with smart pointers.